make block validate + upload code re-usable outside of npm flow
