+++ site/src/lib/api/blocks/shared.ts
@@ -1,1 +1,3 @@
 export const blocksDbCollectionName = "bp-blocks";
+
+export const isRunningOnVercel = !!process.env.VERCEL;
